feat(compose) split bucket into more files and rename to GrapesXXX
